Rename search app to "admin" (Fix #805)
