+++ src/admin/components/JsonData/style.scss
