+++ src/admin/components/SearchResult/index.js
