+++ src/admin/components/SearchResult/style.scss
