+++ src/admin/containers/AddonPage/style.scss
