+++ src/admin/css/SearchResults.scss
