+++ src/admin/css/inc/vars.scss
