+++ src/admin/reducers/search.js
