+++ src/admin/store.js
@@ -4,7 +4,7 @@ import { reducer as reduxAsyncConnect } from 'redux-async-connect';
 import addons from 'core/reducers/addons';
 import api from 'core/reducers/api';
 import auth from 'core/reducers/authentication';
-import search from 'search/reducers/search';
+import search from 'admin/reducers/search';
 import users from 'core/reducers/users';
 import { middleware } from 'core/store';
