+++ tests/client/admin/components/TestJsonData.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { renderIntoDocument, Simulate } from 'react-addons-test-utils';
 import { findDOMNode } from 'react-dom';
 
-import JsonData from 'search/components/JsonData';
+import JsonData from 'admin/components/JsonData';
 
 describe('<JsonData />', () => {
   it('renders an object', () => {
